Process REDCap data in chunks for the EH&S Transfer
Previously, the entire output file from the REDCap dump was loaded
at once. This was causing out of memory errors, as the working memory
grew past 9 GB. In this commit, the REDCap file is read in chunks and
the content of each chunk is processed completely as one batch. Peak
memory consumption as configured is around 3.5 GB.

The original `transform` script took about 51 seconds to process today's
data. This new version took slighly longer at 63 seconds.

I've tested by comparing the output generated by the master version of
transform against this new version. There are no differences in the output.